Initial version of Magento SOAP API v1 client.
Unfortunately libraries that it's using underneath (i.e. xml-build and jquery.soap) are terrible, so it would be good idea to use something else or clean them up.
It would be also a good idea to extract building of Magento SOAP API friendly XML to a separate class/methods at some point.